Forest plots

Create Forest plots consisting of alternating columns of errorbar plots and tabulated numerical statistics.

Usage

forestPlot(
  dfres,
  plotData = NULL,
  plotRelative = TRUE,
  noVar = TRUE,
  referenceParameters = "func",
  sigdigits = 2,
  parameters = unique(dfres$PARAMETER),
  parameterLabels = parameters,
  parameterLabelsPrefix = NULL,
  groupNameLabels = NULL,
  referenceInfo = "auto",
  labelfun = label_value,
  groupname_labelfun = label_value,
  ref_area = c(0.8, 1.25),
  ref_fill_col = "lightgrey",
  ref_fill_alpha = 0.4,
  ref_line_size = 1,
  ref_line_type = "dotted",
  ref_line_col = "black",
  ci_line_type = "solid",
  ci_line_col = "black",
  ci_line_size = 0.7,
  point_shape = 16,
  point_color = "black",
  point_size = 2.5,
  tabTextSize = 10,
  keepYlabs = FALSE,
  keepRightStrip = FALSE,
  stackedPlots = FALSE,
  strip_right_size = NULL,
  strip_top_size = NULL,
  ref_subj_label = "Reference subject",
  ref_area_label = "Reference area",
  point_label = "Point estimate",
  ci_label = "Confidence interval",
  statisticsLabel = "Statistics:",
  xlb = ifelse(plotRelative, "Relative parameter value", "Parameter value"),
  commonXlab = FALSE,
  size.legend.text = rel(0.8),
  return = "plot",
  table = TRUE,
  rightStrip = TRUE,
  errbartabwidth = rep(c(2, 1), length(parameters)),
  errbarplotscale = 1.45,
  tabplotscale = 1.1,
  onlySignificant = FALSE,
  onlySignificantErrorBars = FALSE,
  setSignEff = NULL,
  size = theme_get()$text$size * 0.8,
  addcodeErr = "NULL",
  xlim = c(NA, NA),
  ...
)

Arguments

dfres

A data.frame as output from getForestDFSCM, getForestDFFREM or getForestDFemp.

plotData

A data.frame to be used for creating the Forest plot. If plotData is provided dfres is ignored. See Details.

plotRelative

Should the plot be made on the relative scale (TRUE or FALSE)

noVar

Should the uncertainty in the reference be included (FALSE) or not (TRUE)

referenceParameters

Character string indicating how the reference parameter estimates were obtained. "final" means the final parameter estimates from the model. "func" means the mean parameter estimates across the posterior distribution.

sigdigits

An integer number specifying the number of significant digits to use in the statistics tables.

parameters

A character vector with the parameters present in the dfres$PARAMETER column to be included in the output.

parameterLabels

A vector of labels for the parameters. Should either have the same length as parameters or the same length as the number of rows in dfres. If a vector of the same length as parameters should also be ordered to match parameters. If parameters is not specified, then is should match the order in dfres. Is by default the same as parameters. Used for faceting the errorbar panels in the Forest plot in the x-direction (columns).

parameterLabelsPrefix

A string to be added before the parameterLabels in the facet strip for the parameter panels.

groupNameLabels

A vector of labels for the covariate groups. Should either have the same length as the number unique values in dfres$GROUPNAME or the same length as the number of rows in dfres. Is by default the same as dfres$GROUPNAME.

referenceInfo

A character string or NULL. If set to NULL, no information about how the reference line was derived will be displayed. If set to "auto", generic information based on referenceParameters and the REFROW column in dfres will be included at the bottom of the plot. If not NULL or "auto", then the character string will be displayed below the plot. Uses ggpubr::annotate_figure and ggpubr::text_grob to place and format the text. Formatting instructions to ggpubr::text_grob are passed on through .... Default is "auto".

labelfun

A label function compatible with labeller. Used to format parameterLabels used as column facet labels for the errorbar and table plots. Default is label_value.

groupname_labelfun

A label function compatible with labeller. Used to format groupNameLabels used as row facet labels for the errorbar and table plots. Default is label_value.

ref_area

Numerical vector indicating the horizontal size of the reference area. The default is c(0.8,1.25). Used to multiply the reference value to derive the actual xmin and xmax values for the reference area.

ref_fill_col

Reference area color.

ref_fill_alpha

Reference area alpha value.

ref_line_size

Reference line size.

ref_line_type

Reference line linetype.

ref_line_col

Reference line color.

ci_line_type

Errorbar linetype.

ci_line_col

Errorbar color.

ci_line_size

Errorbar size.

point_shape

Point estimate shape.

point_color

Point estimate color.

point_size

Size of the point estimate symbol.

tabTextSize

The size (pt) of the text in the table plots.

keepYlabs

Logical. Should the labels on the y-axis be kept for all errorbar panels.

keepRightStrip

Logical. Should the right facet titles be kept for all table plots. Only when rightStrip is TRUE.

stackedPlots

Should the plots for the parameters be stacked instead of being vertical. Useful if there are many parameters to visualize. Works best with keepYlab and keepRightStrip set to TRUE.

strip_right_size

Size of the facet text on the right (in pt). Default is NULL, which will fall back on the theme default.

strip_top_size

Size of the facet text on the top (in pt). Default is NULL, which will fall back on the theme default.

ref_subj_label

A character string indicating the label to be used for the reference subject in the figure legend. Default is "Reference subject".

ref_area_label

A character string indicating the label to be used for the reference area in the figure legend. Default is "Reference area".

point_label

A character string indicating the label to be used for the point estimates in the figure legend. Default is "Point estimate".

ci_label

A character string indicating the label to be used for the confidence interval in the figure legend. Default is "Confidence interval".

statisticsLabel

A vector of character string of the same length as parameters to be used as facet labels for the table plots.

xlb

x-axis label for the errorbar plots

commonXlab

Logical. Should a common x-axis title be used or should each panel have its own. Default is FALSE.

size.legend.text

The font size used in the legend in points or relative size to the theme base font. Default is rel(0.8).

return

Either "plot" (default), "plotList" or "data". "plot" returns the ggpubr::ggarrange object. "plotList" returns a list of the different plots that make up the Forest plot (see Details). "data" returns a data.frame with the actual data used to create the Forest plot.

table

Logical. Should the table plots be included in the Forest plot or not.

rightStrip

Should the facet title in the rightmost panel in the Forest plot be displayed? If TRUE (the default) groupNameLabels will be used as the facets labels.

errbartabwidth

A numerical vector of the same length as the number of panels in the plot (eror bar panels + table panels). Specifies the relative width of the panels.

errbarplotscale

Scaling factor for the width of the leftmost errorbar plot to compensate for y-axis labels.

tabplotscale

Scaling factor for the width of the rightmost column (usually a table plot) to adjust for the size of the right strip.

onlySignificant

Logical. Should only the significant covariates be included (TRUE) or all covariates regardless of significance (FALSE).

onlySignificantErrorBars

Logical. Should error bars be hidden for non-significant covariates (TRUE) or be shown for all covariates regardless of significance (FALSE).

addcodeErr

A string of code to be applied to each of the panels with error bars.

...

Arguments passed on to ggpubr::text_grob.

Value

A ggpubr::ggarrange object, a list of plots or a data.frame.

Details

If plotData is NULL, dfres is passed to setupForestPlotData to create a data.frame that contains the exact data to be plotted, including the numerical statistics. If return is set to "data" then the data.frame is returned. This makes it possible to to review the data and/or do modifications if necessary. forestPlot can be called with plotData set to the name of the data.frame, in which case dfres is not required and the plotData data.frame will be used for creating the Forest plot.

If referenceInfo is set to auto (the default) then generic information based on referenceParameters and the REFROW column in dfres will be included at the bottom of the plot. The REFROW column in dfres will be NO if dfRefRow is NULL in the call to getForestDFSCM, getForestDFemp and getForestDFFREM and YES if it is not.

If REFROW is NO and referenceParameters is final then the reference information text will be: "The reference line is based on the final parameter estimates and the reference covariate values in the model.".

If REFROW is NO and referenceParameters is func then the reference information text will be: "The reference line is based on the average parameter estimates over the posterior parameter distribution and the reference covariate values in the model.".

If REFROW is YES and referenceParameters is final then the reference information text will be: "The reference line is based on the final parameter estimates and selected covariate values.".

If REFROW is YES and referenceParameters is func then the reference information text will be: "The reference line is based on the average parameter estimates over the posterior parameter distribution and selected covariate values.".

If referenceInfo is not NULL and not auto then the reference information text will be referenceInfo.

The Forest plots are created as a combination of separate errorbar plots and plots with the table information. Each panel is a separate plot, which are combined using ggpubr::ggarrange. In other words, even if the plot looks like a regular faceted plot it is not. For example, in a one parameter Forest plot with table=TRUE, the left panel is an errorbar plot with the right facet labels suppressed and the right plot is a plot with text, and which have the y-axis labels suppressed. In a plot with three parameters the errobar plot for the middle parameter will have both the y-axis labels and facet labels suppressed. the same is true for that parameter's table plot. The arguments keepYlabs and keepRightStrip controls if the y-axis labels and right facet labels for panels "in the middle" should keep the axis and facet labels or not.

stackedPlots switch from the default stacked horizontal orientation used if multiple parameters are included to vertical stacking. For this to provide a nice display it is necessary to have keepYlabs set to TRUE and keepRightStrip=TRUE. stackedPlots can be combind with table=FALSE and rightStrip=FALSE.

The graphical settings use standard ggplot syntax.

Examples

if (FALSE) {
# Create a Forest plot, including the table part, for all parameters in dfres.

forestPlot(dfres)

# Exclude the table part of the plot

forestPlot(dfres,table=FALSE)

# Create a Forest plot for the specified parameter (which need to be included in dfres).

forestPlot(dfres,parameters="CL")

# Specify group name labels, the size of the table text and x-axis label

forestPlot(dfres,parameters=c("CL"),groupNameLabels = c("Age (y)","Sex","Weight (kg)"),
           tabTextSize = 20,xlb="Relative parameter value")

# Stack the plots instead of plot them horizontally

forestPlot(dfres,parameters = c("CL","Frel"),stackedPlots = TRUE,
           keepYlabs = TRUE,keepRightStrip = TRUE)
}