getForestDFSCM

Get a data frame with Forest border for each univariate or multivariate covariate (and value(s)) in the input data frame. If a list a data frame will be created from the list, see function dfCreateInputForestData

Usage

getForestDFSCM(
  dfCovs,
  cdfCovsNames = NULL,
  functionList = list(function(basethetas, dfrow, ...) {
     return(basethetas[1])
 }),
  functionListName = "PAR1",
  noBaseThetas,
  dfParameters,
  quiet = TRUE,
  probs = c(0.05, 0.95),
  pointFunction = median,
  dfRefRow = NULL,
  cGrouping = NULL,
  ncores = 1,
  cstrPackages = NULL,
  cstrExports = NULL,
  iMiss = -99,
  oneHot = NULL,
  oneHotSep = "_",
  ...
)

Arguments

dfCovs

A data frame with covariates to include, if a covariate value is set -99 or NA, they are assumed missing and will not be included in any FFEM transformations. If a dfCovs is a list an attempt will be made to create the appropriate data frame with the createInputForestData function.

cdfCovsNames

A string vector with names of the rows in dfCovs, if not used, names will be automatically assigned based on the covariate values and column names in dfCovs.

functionList

A list of functions with input (basethetas, covthetas,dfrow and ...) from which the change from the reference value will be calculated. If the function returns a vector of values, each value will be used but functionListName must contain the names with a length of all return for all functions in the functionList

functionListName

A vector of strings (names) of the parameters for each function in the functionList

noBaseThetas

the number of parameters from dfParameters to pass to the functions in functionList.

dfParameters

A data frame with parameter samples from the uncertainty distribution. The vector of final parameter estimates is assumed to be in the first row. The column order is assumed the same as in the NONMEM ext file except the ITERATION and OBJ columns whichshould not be included.

quiet

If output should be allowed during the function call, default= TRUE. (This option is mainly for debugging purposes.)

probs

A vector of probabilities that should be computed for each of the parameters from functionList. These will be used as the as the uncertainties in the Forest plots. The probs vector position one and two will be used for plotting the uncertanties (i.e. columns q1 and q2). Default is c(0.05, 0.95).

pointFunction

The function used to calculate the point for each covariate in the forest plot. default=median This function is also used for the reference covariate combination

dfRefRow

A data frame (one row or equal number of rows as dfCovs) with the covariate values that will be used as the reference, if NULL the typical subject is used as reference.

cGrouping

A vector of numbers defining how to group the y-axis of the Forest plot, the length of the vector should match the number of rows in dfCovs. If NULL (default) an educated guess of the grouping will be set

ncores

the number of cores to use for the calculations, default = 1 which means no parallellization

cstrPackages

a character vector with package names needed to run the calculations in parallel, default = NULL

cstrExports

a character vector with variables needed to run the calculations in parallel, default = NULL

iMiss

The missing value number. -99 by default.

oneHot

An optional one-hot encoding specification (see oneHotEncode). When supplied, raw multi-level categorical columns are one-hot encoded before the parameter functions are evaluated - in dfCovs and dfRefRow for getForestDFSCM(), in dfData for getForestDFemp() - so a single parameter function written against the dummy columns can serve both the parametric and empirical workflows. Default NULL (no encoding). For getForestDFSCM() the raw column is dropped after encoding.

oneHotSep

The separator between the covariate name and the level in the dummy column names created when oneHot is used. Defaults to "_". Set to "" to match dummy columns named like GENO1, GENO2.

...

additional variables to be forwarded to the the functionList functions

Value

A data frame with summary statistics for each parameters and covariate combinations:

See also

Examples

dfData <- read.csv(
  system.file("extdata", "SimVal/DAT-1-MI-PMX-2.csv", package = "PMXForest")
)
extFile <- system.file("extdata", "SimVal/run7.ext", package = "PMXForest")
covFile <- system.file("extdata", "SimVal/run7.cov", package = "PMXForest")

# Covariate rows and uncertainty samples (run7 has 14 THETAs)
dfCovs <- setupDfCovs(dfData,
  covariates = c("WT", "AGE", "CRCL", "FOOD"),
  idVar = "ID"
)
dfSamples <- getSamples(covFile, extFile, n = 50)

# Rebuild the part of the model needed for CL
paramFunction <- function(thetas, df, ...) {
  TVCL <- thetas[4]
  if (any(names(df) == "WT") && df$WT != -99) {
    TVCL <- thetas[4] * (df$WT / 75)^thetas[2]
  }
  if (any(names(df) == "FOOD") && df$FOOD != -99 && df$FOOD == 0) {
    TVCL <- TVCL * (1 + thetas[11])
  }
  list(CL = TVCL)
}

dfres <- getForestDFSCM(
  dfCovs           = dfCovs,
  functionList     = list(paramFunction),
  functionListName = "CL",
  noBaseThetas     = 14,
  dfParameters     = dfSamples
)
head(dfres[, c("COVNAME", "GROUPNAME", "PARAMETER", "POINT", "Q1", "Q2")])
#>     COVNAME GROUPNAME PARAMETER    POINT        Q1       Q2
#> 1   WT=58.5        WT        CL  9.88298  9.454588 10.41761
#> 2    WT=126        WT        CL 17.57110 16.809454 18.52163
#> 3    AGE=22       AGE        CL 11.90740 11.391257 12.55155
#> 4    AGE=65       AGE        CL 11.90740 11.391257 12.55155
#> 5 CRCL=74.9      CRCL        CL 11.90740 11.391257 12.55155
#> 6  CRCL=150      CRCL        CL 11.90740 11.391257 12.55155