getSamples

Function that gets samples, either from a regular bootstrap (Psn output) a SIR (PsN output), a covariance matrix (NM output) or samples from a small bootstrap using the estimated covariance matrix

Usage

getSamples(input, extFile = NULL, n = NULL, indexvec = NULL, zerosindex = NULL)

Arguments

input

The input file name (character string). If the file name extension is `.cov` it is assumed that the file is a a NONMEM .cov file and if the extension is `.csv` is is assumed that it is a PsN raw results file from either a bootstrap or a SIR. Can also be a data.frame with the same structure as such files.

extFile

The name of the NONMEM ext file. Needed to transform output to the ext-format. Should not be NULL.

n

The number of samples to generate when sampling is done from a variance-covaraince matrix. NULL is permissable if the input is a SIR or bootstrap raw_results file.

indexvec

A vector that can be used to subset the columns in the raw results file so that the columns comes in the order THETA, SIGMA, OMEGA. not required if a PsN raw_results_structure is present in the same directory as the csv file.

zerosindex

A vector with indicies that indicates which parameters that are fixed to zero.

Value

A data frame with parameter vectors.

In case a .cov file is provided as input the parameter vectors is the result of multivariate sampling from the variance-covariance matrix.

In case a .csv file is provided as input and n=NULL, the input is assumed to be a SIR or bootstrap file. In case of a SIR file, all samples present in the .csv file is returned. In case it is a bootstrap file, the function returns all samples with a nonzero value in the ofv column.

In case a .csv file is provided as input and n is not NULL, it is assumed that the raw results come from a boostrap analysis. A variance- covariance matrix will be created from the samples with a non-zero value in the ofv column and n parameter vectors will be returned obtained by multivariate sampling from the constructed variance-covariance matrix.

Examples

if (FALSE) {

# Covariance matrix
getSamples(covFile,extFile,n=175)

# Bootstrap
getSamples(BSFile,extFile)

# SIRFile
getSamples(SIRFile,extFile)

# Small bootstrap mvnorm sampling
getSamples(BSFile,extFile,n=175)

# Data frame
dfParameters<-read.csv(rawresFile)

getSamples(dfParameters)

# Data frame mvnorm sampling
getSamples(dfParameters,n=175)
}