Set up the data.frame for forestPlot to use

Creates the data.frame that is actually used for generating the Forest plots. Used by forestPlot.

Usage

setupForestPlotData(
  dfres,
  parameters = unique(dfres$PARAMETER),
  parameterLabels = NULL,
  parameterLabelsPrefix = NULL,
  groupNameLabels = NULL,
  statisticsLabels = NULL,
  plotRelative = TRUE,
  noVar = FALSE,
  reference = "func",
  sigdigits = 2,
  onlySignificant = FALSE,
  setSignEff = NULL
)

Arguments

dfres

A data.frame as output from getForestDFSCM, getForestDFFREM or getForestDFemp.

parameters

A character vector with the parameters present in the dfres$PARAMETER column to be included in the output.

parameterLabels

A vector of labels for the parameters. Should either have the same length as parameters or the same length as the number of rows in dfres. If a vector of the same length as parameters should also be ordered to match parameters. If parameters is not specified, then is should match the order in dfres. Is by default the same as parameters. Used for faceting the errorbar panels in the Forest plot in the x-direction (columns).

parameterLabelsPrefix

A string to be added before the parameterLabels in the facet strip for the parameter panels.

groupNameLabels

A vector of labels for the covariate groups. Should either have the same length as the number unique values in dfres$GROUPNAME or the same length as the number of rows in dfres. Is by default the same as dfres$GROUPNAME.

statisticsLabels

A character string that will precede the parameterLabels in the facet labels for the statistics panels. Default is `Statistics:`.

plotRelative

Should the plot be made on the relative scale (TRUE or FALSE)

noVar

Should the uncertainty in the reference be included (FALSE) or not (TRUE)

reference

Which reference should be used, the final parameter estimates (final) or the parameter function (func)

sigdigits

An integer number specifying the number of significant digits to use in the statistics tables.

onlySignificant

Logical. Should only the significant covariates be included (TRUE) or all covariates regardless of significance (FALSE).

Value

A processed data.frame to the used for creating the Forest plot. Only the columns used in the actual Forest plot is included:

GROUPNAME

The names of the covariate group, as specified in the call to createInputForestData.

GROUPNAMELABEL

The names of the covariate group as they will appear in the y-axis Facet. Is by default the same as GROUPNAME.

COVNAME

The label for each row in each facet in the Forest plot. For exampe Males and Females for the GROUPNAME Sex.

PARAMETER

The parameters to be plotted in the Forest plot. Comes from functionListName in the call to getForestDFSCM.

PARAMETERLABEL

The labels for PARAMETER as they will appear in the Forest plot. Is by default the same as PARAMETER. Used for faceting the errorbar panels in the Forest plot in the x-direction (columns).

REF

The parameter value to be used for the reference line, either REFTRUE or REFFUNC in dfres or 1 if plotRelative=TRUE.

point

The value used to plot the point estimate for each COVNAME.

q1 and q2

The values used to plot the lower (q1) and (q2) interval edges of the confidence intervals.

reference

A character string indicating what was used as the reference, either func or true, for the parameter function or the first row in dfParameters

STATISTIC

A character variable with the numerical summary for each COVNAME to be used in the statistics table.

STATISTICSLABEL

Similar to PARAMETERLABEL. Is the label used for PARAMETER in the table plots. Used for faceting the table panels in the Forest plot in the x-direction (columns).

onlySignificant

Logical. Should only GROUPNAME with with TRUE in the COVEFF column be included in the Forest plot?

setSignEff

NULL of a list. If a list, it should be a list ov vectors with a string for PARAMETER as the forst element and a string for GROUPNAME as the second element. If they match PARAMETER and GROUPNAME in dfres then COVEFF will be set to TRUE else be set to FALSE. This will override the existing values in COVEFF.

Examples

if (FALSE) {
plotData<- setupForestPlotData(dfres)
}